Move Date & Time to Events and Maps to a new Venues tab, and split class-venue.php into Venue instance + Venue_Setup singleton.
