+++ includes/core/classes/class-event.php
@@ -546,7 +546,7 @@ class Event {
 
 		if ( ! empty( $term ) && is_a( $term, 'WP_Term' ) ) {
 			$venue_information['name'] = $term->name;
-			$venue                     = Venue::get_instance()->get_venue_post_from_term_slug( $term->slug );
+			$venue                     = Venue_Setup::get_instance()->get_venue_post_from_term_slug( $term->slug );
 
 			if ( 'online-event' === $term->slug ) {
 				$venue_information['is_online_event'] = true;
