+++ includes/core/classes/class-setup.php
@@ -74,13 +74,13 @@ class Setup {
 		Settings::get_instance();
 		Settings\Credits::get_instance();
 		Settings\Events::get_instance();
-		Settings\Formatting::get_instance();
 		Settings\Roles::get_instance();
 		Settings\Rsvp_Settings::get_instance();
 		Settings\Tools::get_instance();
+		Settings\Venues::get_instance();
 		Topic::get_instance();
 		User::get_instance();
-		Venue::get_instance();
+		Venue_Setup::get_instance();
 	}
 
 	/**
@@ -345,7 +345,7 @@ class Setup {
 	 * @return void
 	 */
 	public function add_online_event_term(): void {
-		Venue::get_instance()->register_taxonomy();
+		Venue_Setup::get_instance()->register_taxonomy();
 
 		$term_name = __( 'Online event', 'gatherpress' );
 		$term_slug = 'online-event';
