+++ test/unit/php/includes/tests/core/classes/blocks/class-test-venue.php
@@ -11,6 +11,7 @@ namespace GatherPress\Tests\Core\Blocks;
 use GatherPress\Core\Blocks\Venue as Venue_Block;
 use GatherPress\Core\Event;
 use GatherPress\Core\Venue;
+use GatherPress\Core\Venue_Setup;
 use GatherPress\Tests\Base;
 use WP_Block;
 
@@ -691,7 +692,7 @@ class Test_Venue extends Base {
 		$venue_post = get_post( $venue_id );
 
 		// Create the venue term with the correct slug format.
-		$term_slug = Venue::get_instance()->get_venue_term_slug( $venue_post->post_name );
+		$term_slug = Venue_Setup::get_instance()->get_venue_term_slug( $venue_post->post_name );
 		wp_insert_term(
 			'Linked Venue For Event Test',
 			Venue::TAXONOMY,
@@ -985,7 +986,7 @@ class Test_Venue extends Base {
 		$venue_post = get_post( $venue_id );
 
 		// Create the venue term with the correct slug format.
-		$term_slug = Venue::get_instance()->get_venue_term_slug( $venue_post->post_name );
+		$term_slug = Venue_Setup::get_instance()->get_venue_term_slug( $venue_post->post_name );
 		wp_insert_term(
 			'Override Venue Test',
 			Venue::TAXONOMY,
